Split dateTime utils to not require config
