+++ packages/shared/src/utils/dateTime.js
@@ -11,13 +11,11 @@ import {
   sub,
   startOfDay,
 } from 'date-fns';
-import { formatInTimeZone } from 'date-fns-tz';
-import config from 'config';
 import { TIME_UNIT_OPTIONS } from '@tamanu/constants';
 
-const ISO9075_DATE_FORMAT = 'yyyy-MM-dd';
-const ISO9075_DATETIME_FORMAT = 'yyyy-MM-dd HH:mm:ss';
-const ISO8061_WITH_TIMEZONE = "yyyy-MM-dd'T'HH:mm:ssXXX";
+export const ISO9075_DATE_FORMAT = 'yyyy-MM-dd';
+export const ISO9075_DATETIME_FORMAT = 'yyyy-MM-dd HH:mm:ss';
+export const ISO8061_WITH_TIMEZONE = "yyyy-MM-dd'T'HH:mm:ssXXX";
 
 export const isISOString = dateString =>
   isMatch(dateString, ISO9075_DATETIME_FORMAT) || isMatch(dateString, ISO9075_DATE_FORMAT);
@@ -64,48 +62,6 @@ export function toDateString(date) {
   }
   const dateObj = parseDate(date);
   return formatISO9075(dateObj, { representation: 'date' });
-}
-
-// CountryDateTime functions are server only
-// Servers require a specific reference to timeZone since most of our servers are in UTC
-export function toCountryDateTimeString(date) {
-  if (date === null || date === undefined) {
-    return null;
-  }
-
-  return formatInTimeZone(date, config?.countryTimeZone, ISO9075_DATETIME_FORMAT);
-}
-
-export function toCountryDateString(date) {
-  if (date === null || date === undefined) {
-    return null;
-  }
-
-  return formatInTimeZone(date, config?.countryTimeZone, ISO9075_DATE_FORMAT);
-}
-
-export function dateTimeStringIntoCountryTimezone(date) {
-  if (date === null || date === undefined) {
-    return null;
-  }
-
-  return parseISO(formatInTimeZone(date, config?.countryTimeZone, ISO8061_WITH_TIMEZONE));
-}
-
-export function getCurrentCountryTimeZoneDateTimeString() {
-  // Use the countryTimeZone if set, other wise fallback to the server time zone
-  if (config?.countryTimeZone) {
-    return formatInTimeZone(new Date(), config.countryTimeZone, ISO9075_DATETIME_FORMAT);
-  }
-  return formatISO9075(new Date());
-}
-
-export function getCurrentCountryTimeZoneDateString() {
-  // Use the countryTimeZone if set, other wise fallback to the server time zone
-  if (config?.countryTimeZone) {
-    return formatInTimeZone(new Date(), config.countryTimeZone, ISO9075_DATE_FORMAT);
-  }
-  return formatISO9075(new Date(), { representation: 'date' });
 }
 
 export function getCurrentDateTimeString() {
